refactor(media): Add services/types.ts
